+++ tests/test_generate_summary_report.py
@@ -1,6 +1,7 @@
 """Test the view.summary_report.generate_summary_report function"""
 
 import logging
+import pathlib
 
 import pytest
 
@@ -8,22 +9,20 @@ import view.summary_report
 
 
 @pytest.mark.parametrize("output_file_name_base", ["test_1", "test_2"])
-def test_generate_summary_report(output_file_name_base: str, test_data_directory: str) -> None:
+def test_generate_summary_report(output_file_name_base: str, tmp_path: pathlib.Path) -> None:
     """Test the view.summary_report.generate_summary_report function
 
     Parameters:
     output_file_name_base (str): The file name base to use when creating the report
-    test_data_directory (str): The path to the temporary test data directory
+    tmp_path (pathlib.Path): The path used to stage files needed for testing (pytest fixture)
 
     Returns:
     None
     """
-    output_directory = f"{test_data_directory}/reports"
+    output_path = tmp_path / f"{output_file_name_base}.html"
 
     # Generate the report
-    view.summary_report.generate_summary_report(output_directory=output_directory,
-        output_file_name=output_file_name_base)
-    output_path = f"{output_directory}/{output_file_name_base}.html"
+    view.summary_report.generate_summary_report(output_path=output_path)
 
     # Test the generated report
     with open(output_path, "r", encoding="utf_8") as report_file: