+++ view/summary_report.py
@@ -1,23 +1,73 @@
 """Generate a summary HTML report for RimWorld save game data"""
 
+import pathlib
+
 import dominate
-from dominate.tags import attr, div, h1, h2, li, link, p, ul
+from dominate.util import raw
+from dominate.tags import attr, div, h1, h2, h3, li, link, p, ul
+import pandas
+import plotly.express
 
 import extract.extract_save_data
 
 
-def generate_summary_report(output_directory: str, output_file_name: str) -> None:
+def get_environment_section(pawn_data: list, weather_data: dict) -> None:
+    """Build the environment and weather section of the report
+
+    Parameters:
+    pawn_data (list): The list of dictionaries containing pawn data
+    weather_data (dict): A dictionary containing weather data for the map being analyzed
+
+    Returns:
+    None
+    """
+    h2("Environment and Weather")
+    h3("Weather")
+
+    with ul():
+        li(f"Current weather: {weather_data['weather_current']}")
+        li(f"Last weather: {weather_data['weather_last']}")
+        li(f"Current weather age: {weather_data['weather_current_age']}")
+
+    h3("Pawn Ambient Temperatures")
+
+    with ul():
+        for pawn in pawn_data:
+            ambient_temperature = round(float(pawn['pawn_ambient_temperature']), 1)
+            ambient_temperature_string = f"{ambient_temperature}&#176;C"
+            pawn_temperature_string = f"{ambient_temperature_string} temperature felt by \
+                {pawn['pawn_name_full']}"
+            li(raw(pawn_temperature_string))
+
+
+def get_histogram_html(dataframe: pandas.core.frame.DataFrame, x_axis_field: str,
+    labels: dict) -> str:
+    """Return the HTML for a histogram chart
+
+    Parameters:
+    dataframe (pandas.core.frame.DataFrame): The pandas DataFrame to use in the chart
+    x_axis_field (str): The field to use for the x-axis series
+    labels (dict): A dictionary with the chart labels metadata
+
+    Returns:
+    str: A histogram chart HTML
+    """
+    fig = plotly.express.histogram(dataframe, x=x_axis_field, labels=labels)
+    fig.update_layout(bargap=0.05, yaxis_title_text="Count")
+
+    return fig.to_html(full_html=False)
+
+
+def generate_summary_report(output_path: pathlib.Path) -> None:
     """Generate an HTML report with a list of the installed mods found
 
     Parameters:
-    output_directory (str): The directory to store the generated report
-    output_file_name (str): The name to use for the report HTML file
+    output_path (pathlib.Path): The file path where the report should be created
 
     Returns:
     None
     """
     mod_list = extract.extract_save_data.extract_mod_list()
-    output_path = f"{output_directory}/{output_file_name}.html"
     doc = dominate.document(title='RimWorld Save Game Summary Report')
 
     with doc.head:
@@ -50,11 +100,12 @@ def generate_summary_report(output_directory: str, output_file_name: str) -> Non
                     li(f"{pawn['pawn_name_full']}, age {pawn['pawn_biological_age']}")
 
             h2(f"Plants ({extract.extract_save_data.get_plant_count()})")
+            plant_data_raw = extract.extract_save_data.get_plant_data()
 
             with ul():
                 displayed_plant_types = []
 
-                for plant in extract.extract_save_data.get_plant_data():
+                for plant in plant_data_raw:
                     if plant['plant_definition'] in displayed_plant_types:
                         continue
 
@@ -68,6 +119,24 @@ def generate_summary_report(output_directory: str, output_file_name: str) -> Non
 
                     if len(displayed_plant_types) >= 20:
                         break
+
+            plant_dataframe = extract.extract_save_data.plant_dataframe(dictionary_list=\
+                plant_data_raw)
+            p(raw(plant_dataframe.head().to_html()))
+            p(raw(plant_dataframe.tail().to_html()))
+            p(raw(plant_dataframe.describe().to_html()))
+            p(raw(plant_dataframe["plant_definition"].value_counts().to_frame().to_html()))
+            raw(
+                get_histogram_html(
+                    dataframe=plant_dataframe,
+                    x_axis_field="plant_growth_bin",
+                    labels={"plant_growth_bin": "Plant growth (%)"}
+                )
+            )
+            get_environment_section(
+                pawn_data=extract.extract_save_data.get_pawn_data(),
+                weather_data=extract.extract_save_data.get_weather_data()
+            )
 
     with open(output_path, "w", encoding="utf_8") as output_file:
         output_file.write(str(doc))